requirements.txt
Flask>=3.0.0
python-chess>=1.10.0
opencv-python>=4.10.0
numpy>=1.26.0
Pillow>=10.0.0
stockfish>=3.28.0
